+++ src/CreateNewDoctorModel.java
@@ -6,7 +6,7 @@ import java.util.Map;
 /**handles all calculations, database queries, and the overall work needed to be done for handling this associated role
  * Does NOT interact with the view class directly, and also does NOT interact with the Controller class(The Controller
  * class interacts with this class, not the other way around.)*/
-public class CreateNewDoctorModel {
+public class CreateNewDoctorModel extends UserSuperClass {
 	
 	private String name;
 	private String username;
@@ -126,16 +126,7 @@ public class CreateNewDoctorModel {
 		return temp.toArray(new String[0]);
 	}
 
-	
-	//return the department list in a string[] format for use with combobox
-	public String[] getDeptList() {
-		
-		ArrayList<String> temp = Main.dbaseClass.getDepartmentList();
-		
-		
-		return temp.toArray(new String[0]);
-	}
-	
+
 	
 /**Getters and Setters*/